Fix the methods involved with wiki->md conversion.
The original version from zweifisch didn't work anymore, at least
not with python 3.8.1. I needed that functionality and made it work.
Good enough for my purpose. It doesn't convert links between both
versions but that was ok for me.

Also black formatted a good chunk of this code so it's not obvious
which parts of the old code didn't work in a current python version.
Sorry for that.